requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
requests==2.31.0
spotipy==2.23.0
python-dotenv==1.0.0
APScheduler==3.10.4
PyJWT==2.8.0
yandex-music==2.1.0